Добавил хранимую процедуру, исправил баг в заглушке SetStatusTrack

+++ src/OwnRadio.Server.Core/OwnRadio.Web.Api/src/OwnRadio.Web.Api/Models/Track.cs
@@ -85,7 +85,7 @@ namespace OwnRadio.Web.Api.Models
 				// Открываем соединение
 				npgSqlConnection.Open();
 				// Выполняем хранимую процедуру (функцию)
-				rowCount = (int)npgSqlCommand.ExecuteNonQuery();
+				rowCount = npgSqlCommand.ExecuteNonQuery();
 			}
 			return rowCount;
 		}
